Fixed P300 target indication - merged stimulus presenter collection target indicator class with p300 command centre to resolve referencing issues
