+++ Runtime/Scripts/Behaviors/Training/TrainingInterfaces.cs
@@ -1,0 +1,17 @@
+namespace BCIEssentials
+{
+    public interface ITargetIndicator
+    {
+        public int TargetCount { get; }
+
+        public void BeginTargetIndication(int index);
+        public void EndTargetIndication();
+    }
+
+
+    public interface ITargetable
+    {
+        public void StartTargetIndication();
+        public void EndTargetIndication();
+    }
+}
